+++ internal/cmd/start_orphan_windows.go
@@ -1,0 +1,16 @@
+//go:build windows
+
+package cmd
+
+import (
+	"fmt"
+
+	"github.com/steveyegge/gastown/internal/style"
+)
+
+// cleanupOrphanedClaude is a Windows stub.
+// Orphan cleanup requires Unix-specific signals (SIGTERM/SIGKILL).
+func cleanupOrphanedClaude(graceSecs int) {
+	fmt.Printf("  %s Orphan cleanup not supported on Windows\n",
+		style.Dim.Render("○"))
+}
